+++ app_runner/src/app_runner/specs/inputs_spec.py
@@ -13,6 +13,7 @@ from app_runner.specs.inputs.file_scp_spec import FileScpSpec
 
 if TYPE_CHECKING:
     from pathlib import Path
+    from bfabric import Bfabric
 
 InputSpecType = Annotated[
     BfabricResourceSpec | FileScpSpec | BfabricDatasetSpec | BfabricOrderFastaSpec | BfabricAnnotationSpec,
@@ -25,7 +26,11 @@ class InputsSpec(BaseModel):
     inputs: list[InputSpecType]
 
     @classmethod
-    def read_yaml(cls, path: Path) -> list[InputSpecType]:
+    def read_yaml(cls, path: Path) -> InputsSpec:
+        return cls.model_validate(yaml.safe_load(path.read_text()))
+
+    @classmethod
+    def read_yaml_old(cls, path: Path) -> list[InputSpecType]:
         model = cls.model_validate(yaml.safe_load(path.read_text()))
         return model.inputs
 
@@ -33,3 +38,10 @@ class InputsSpec(BaseModel):
     def write_yaml(cls, specs: list[InputSpecType], path: Path) -> None:
         model = cls.model_validate(dict(specs=specs))
         path.write_text(yaml.dump(model.model_dump(mode="json")))
+
+    def apply_filter(self, filter: str, client: Bfabric) -> InputsSpec:
+        matches = []
+        for spec in self.inputs:
+            if spec.resolve_filename(client) == filter:
+                matches.append(spec)
+        return type(self)(inputs=matches)