+++ app_runner/docs/changelog.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - New input type `bfabric_order_fasta` which will place an order fasta file to the specified path, or create an empty
     file if there is no order fasta available.
+- `--filter` flag has been added to `inputs prepare` and `inputs clean` commands.
 
 ## \[0.0.11\] - 2025-01-16